+++ griptape/drivers/prompt/amazon_bedrock_prompt_driver.py
@@ -1,15 +1,19 @@
 from __future__ import annotations
-from typing import TYPE_CHECKING, Any
+
 from collections.abc import Iterator
-from attrs import define, field, Factory
-from griptape.drivers import BasePromptDriver
+from typing import TYPE_CHECKING, Any
+
+from attrs import Factory, define, field
+
 from griptape.artifacts import TextArtifact
+from griptape.drivers import BasePromptDriver
+from griptape.tokenizers import AmazonBedrockTokenizer, BaseTokenizer
 from griptape.utils import import_optional_dependency
-from griptape.tokenizers import SimpleTokenizer, BaseTokenizer
 
 if TYPE_CHECKING:
-    from griptape.utils import PromptStack
     import boto3
+
+    from griptape.utils import PromptStack
 
 
 @define
@@ -19,7 +23,9 @@ class AmazonBedrockPromptDriver(BasePromptDriver):
         default=Factory(lambda self: self.session.client("bedrock-runtime"), takes_self=True), kw_only=True
     )
     additional_model_request_fields: dict = field(default=Factory(dict), kw_only=True)
-    tokenizer: BaseTokenizer = field(default=Factory(lambda: SimpleTokenizer(characters_per_token=4)), kw_only=True)
+    tokenizer: BaseTokenizer = field(
+        default=Factory(lambda self: AmazonBedrockTokenizer(model=self.model), takes_self=True), kw_only=True
+    )
 
     def try_run(self, prompt_stack: PromptStack) -> TextArtifact:
         response = self.bedrock_client.converse(**self._base_params(prompt_stack))
@@ -42,10 +48,12 @@ class AmazonBedrockPromptDriver(BasePromptDriver):
 
     def _base_params(self, prompt_stack: PromptStack) -> dict:
         system_messages = [
-            {"text": input.content} for input in prompt_stack.inputs if input.is_system() and input.content
+            self.tokenizer.prompt_stack_input_to_message(input)
+            for input in prompt_stack.inputs
+            if input.is_system() and input.content
         ]
         messages = [
-            {"role": self.__to_amazon_bedrock_role(input), "content": [{"text": input.content}]}
+            self.tokenizer.prompt_stack_input_to_message(input)
             for input in prompt_stack.inputs
             if not input.is_system()
         ]
@@ -57,11 +65,3 @@ class AmazonBedrockPromptDriver(BasePromptDriver):
             "inferenceConfig": {"temperature": self.temperature},
             "additionalModelRequestFields": self.additional_model_request_fields,
         }
-
-    def __to_amazon_bedrock_role(self, prompt_input: PromptStack.Input) -> str:
-        if prompt_input.is_system():
-            return "system"
-        elif prompt_input.is_assistant():
-            return "assistant"
-        else:
-            return "user"