+++ griptape/drivers/prompt/base_prompt_driver.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 from abc import ABC, abstractmethod
-from typing import TYPE_CHECKING, Optional, Callable
+from typing import TYPE_CHECKING, Optional
 from collections.abc import Iterator
 from attrs import define, field, Factory
 from griptape.events import StartPromptEvent, FinishPromptEvent, CompletionChunkEvent
@@ -32,9 +32,6 @@ class BasePromptDriver(SerializableMixin, ExponentialBackoffMixin, ABC):
     temperature: float = field(default=0.1, kw_only=True, metadata={"serializable": True})
     max_tokens: Optional[int] = field(default=None, kw_only=True, metadata={"serializable": True})
     structure: Optional[Structure] = field(default=None, kw_only=True)
-    prompt_stack_to_string: Callable[[PromptStack], str] = field(
-        default=Factory(lambda self: self.default_prompt_stack_to_string_converter, takes_self=True), kw_only=True
-    )
     ignored_exception_types: tuple[type[Exception], ...] = field(
         default=Factory(lambda: (ImportError, ValueError)), kw_only=True
     )
@@ -42,32 +39,23 @@ class BasePromptDriver(SerializableMixin, ExponentialBackoffMixin, ABC):
     tokenizer: BaseTokenizer
     stream: bool = field(default=False, kw_only=True, metadata={"serializable": True})
 
-    def max_output_tokens(self, text: str | list) -> int:
-        tokens_left = self.tokenizer.count_output_tokens_left(text)
-
-        if self.max_tokens:
-            return min(self.max_tokens, tokens_left)
-        else:
-            return tokens_left
-
-    def token_count(self, prompt_stack: PromptStack) -> int:
-        return self.tokenizer.count_tokens(self.prompt_stack_to_string(prompt_stack))
-
     def before_run(self, prompt_stack: PromptStack) -> None:
         if self.structure:
             self.structure.publish_event(
                 StartPromptEvent(
                     model=self.model,
-                    token_count=self.token_count(prompt_stack),
+                    token_count=self.tokenizer.count_tokens(prompt_stack),
                     prompt_stack=prompt_stack,
-                    prompt=self.prompt_stack_to_string(prompt_stack),
+                    prompt=self.tokenizer.prompt_stack_to_string(prompt_stack),
                 )
             )
 
     def after_run(self, result: TextArtifact) -> None:
         if self.structure:
             self.structure.publish_event(
-                FinishPromptEvent(model=self.model, token_count=result.token_count(self.tokenizer), result=result.value)
+                FinishPromptEvent(
+                    model=self.model, result=result.value, token_count=self.tokenizer.count_tokens(result.value)
+                )
             )
 
     def run(self, prompt_stack: PromptStack) -> TextArtifact:
@@ -91,21 +79,6 @@ class BasePromptDriver(SerializableMixin, ExponentialBackoffMixin, ABC):
                 return result
         else:
             raise Exception("prompt driver failed after all retry attempts")
-
-    def default_prompt_stack_to_string_converter(self, prompt_stack: PromptStack) -> str:
-        prompt_lines = []
-
-        for i in prompt_stack.inputs:
-            if i.is_user():
-                prompt_lines.append(f"User: {i.content}")
-            elif i.is_assistant():
-                prompt_lines.append(f"Assistant: {i.content}")
-            else:
-                prompt_lines.append(i.content)
-
-        prompt_lines.append("Assistant:")
-
-        return "\n\n".join(prompt_lines)
 
     @abstractmethod
     def try_run(self, prompt_stack: PromptStack) -> TextArtifact: ...