+++ griptape/drivers/prompt_model/base_prompt_model_driver.py
@@ -11,7 +11,6 @@ from griptape.mixins import SerializableMixin
 
 @define
 class BasePromptModelDriver(SerializableMixin, ABC):
-    max_tokens: Optional[int] = field(default=None, kw_only=True)
     prompt_driver: Optional[BasePromptDriver] = field(default=None, kw_only=True)
     supports_streaming: bool = field(default=True, kw_only=True)