+++ griptape/tokenizers/simple_tokenizer.py
@@ -7,14 +7,9 @@ from griptape.tokenizers import BaseTokenizer
 @define()
 class SimpleTokenizer(BaseTokenizer):
     model: Optional[str] = field(init=False, kw_only=True, default=None)
-    max_input_tokens: int = field(kw_only=True, default=0)
-    max_output_tokens: int = field(kw_only=True, default=0)
     characters_per_token: int = field(kw_only=True)
 
-    def count_tokens(self, text: str | list) -> int:
-        if isinstance(text, str):
-            num_tokens = (len(text) + self.characters_per_token - 1) // self.characters_per_token
+    def try_count_tokens(self, text: str) -> int:
+        num_tokens = (len(text) + self.characters_per_token - 1) // self.characters_per_token
 
-            return num_tokens
-        else:
-            raise ValueError("Text must be a string.")
+        return num_tokens