+++ tests/unit/tokenizers/test_hugging_face_tokenizer.py
@@ -2,9 +2,10 @@ from os import environ
 
 environ["TRANSFORMERS_VERBOSITY"] = "error"
 
-import pytest
-from transformers import GPT2Tokenizer
-from griptape.tokenizers import HuggingFaceTokenizer
+import pytest  # noqa: E402
+from griptape.utils import PromptStack  # noqa: E402
+from transformers import GPT2Tokenizer  # noqa: E402
+from griptape.tokenizers import HuggingFaceTokenizer  # noqa: E402
 
 
 class TestHuggingFaceTokenizer:
@@ -13,6 +14,9 @@ class TestHuggingFaceTokenizer:
         return HuggingFaceTokenizer(tokenizer=GPT2Tokenizer.from_pretrained("gpt2"), max_output_tokens=1024)
 
     def test_token_count(self, tokenizer):
+        assert (
+            tokenizer.count_tokens(PromptStack(inputs=[PromptStack.Input(content="foo bar huzzah", role="user")])) == 6
+        )
         assert tokenizer.count_tokens("foo bar huzzah") == 5
 
     def test_input_tokens_left(self, tokenizer):